requirements.txt
mediapipe>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
pyttsx3>=2.90
